Split up Stan wrapper module into two submodules
This is because BridgeStan is not Nix-packaged yet and thus would
make for a missing import if a user uses the Nix derivation in
`examples/nix/chainsail_helpers.nix` and then only wants to use the
httpstan wrapper.
