+++ examples/stan-mixture/README.md
@@ -1,3 +1,3 @@
 # Stan Gaussian mixture
 This is an example of how to wrap a model defined in the [Stan](https://mc-stan.org) language for use with Chainsail.
-Chainsail internally has a [httpstan](https://github.com/stan-dev/httpstan) server running that is being accessed by the Stan PDF wrapper defined in `chainsail_utils.pdf.stan`. 
+Chainsail internally has a [httpstan](https://github.com/stan-dev/httpstan) server running that is being accessed by the Stan PDF wrapper defined in `chainsail_utils.pdf.stan.httpstan`.
